+++ cmd/entire/cli/agent/skill_events.go
@@ -1,69 +1,32 @@
 package agent
 
-// Skill event types recorded in checkpoint metadata.
+import "github.com/entireio/cli/cmd/entire/cli/agent/types"
+
+// The skill-event types and their constants live in the leaf agent/types
+// package so the checkpoint contract can construct and reference skill events
+// without importing the full agent package. These aliases keep existing
+// agent.SkillEvent* references working.
 const (
-	SkillEventTypePromptInvocation = "prompt_invocation"
-	SkillEventTypeToolInvocation   = "tool_invocation"
+	SkillEventTypePromptInvocation = types.SkillEventTypePromptInvocation
+	SkillEventTypeToolInvocation   = types.SkillEventTypeToolInvocation
+
+	SkillSignalPiInputSlashCommand = types.SkillSignalPiInputSlashCommand
+	SkillSignalPromptSlashCommand  = types.SkillSignalPromptSlashCommand
+	SkillSignalClaudeSkillToolUse  = types.SkillSignalClaudeSkillToolUse
+
+	SkillConfidenceExplicit = types.SkillConfidenceExplicit
+
+	SkillCollapseTargetUserMessage = types.SkillCollapseTargetUserMessage
+	SkillCollapseTargetToolPair    = types.SkillCollapseTargetToolPair
 )
 
-// Skill event source signals.
-const (
-	SkillSignalPiInputSlashCommand = "input_slash_command"
-	SkillSignalPromptSlashCommand  = "prompt_slash_command"
-	SkillSignalClaudeSkillToolUse  = "skill_tool_use"
+type (
+	SkillEvent                 = types.SkillEvent
+	SkillEventSkill            = types.SkillEventSkill
+	SkillEventSource           = types.SkillEventSource
+	SkillEventTranscriptAnchor = types.SkillEventTranscriptAnchor
+	SkillEventCollapse         = types.SkillEventCollapse
 )
-
-// Skill event confidence values.
-const (
-	SkillConfidenceExplicit = "explicit"
-)
-
-// Skill event collapse targets.
-const (
-	SkillCollapseTargetUserMessage = "user_message"
-	SkillCollapseTargetToolPair    = "tool_pair"
-)
-
-// SkillEvent records a native agent skill signal without rewriting the raw transcript.
-// Consumers use TranscriptAnchor/Native to locate the underlying raw event and Collapse
-// to decide whether/how to hide verbose skill material by default.
-type SkillEvent struct {
-	ID        string           `json:"id,omitempty"`
-	EventType string           `json:"event_type"`
-	Skill     SkillEventSkill  `json:"skill"`
-	Source    SkillEventSource `json:"source"`
-
-	TurnID    string `json:"turn_id,omitempty"`
-	Timestamp string `json:"timestamp,omitempty"`
-
-	TranscriptAnchor *SkillEventTranscriptAnchor `json:"transcript_anchor,omitempty"`
-	Native           map[string]string           `json:"native,omitempty"`
-	Collapse         SkillEventCollapse          `json:"collapse"`
-}
-
-type SkillEventSkill struct {
-	Name string `json:"name"`
-}
-
-type SkillEventSource struct {
-	Agent      string `json:"agent"`
-	Signal     string `json:"signal"`
-	Confidence string `json:"confidence"`
-}
-
-type SkillEventTranscriptAnchor struct {
-	Unit      string   `json:"unit,omitempty"`
-	Start     int      `json:"start,omitempty"`
-	End       int      `json:"end,omitempty"`
-	EntryIDs  []string `json:"entry_ids,omitempty"`
-	ToolUseID string   `json:"tool_use_id,omitempty"`
-}
-
-type SkillEventCollapse struct {
-	Target           string `json:"target"`
-	Label            string `json:"label,omitempty"`
-	DefaultCollapsed bool   `json:"default_collapsed"`
-}
 
 // SkillEventExtractor is implemented by agents that can derive native skill events
 // from their transcript format.
