+++ cmd/entire/cli/agent/types.go
@@ -1,6 +1,10 @@
 package agent
 
-import "time"
+import (
+	"time"
+
+	"github.com/entireio/cli/cmd/entire/cli/agent/types"
+)
 
 // HookType represents agent lifecycle events
 type HookType string
@@ -43,19 +47,7 @@ type SessionChange struct {
 	Timestamp  time.Time
 }
 
-// TokenUsage represents aggregated token usage for a checkpoint.
-// This is agent-agnostic and can be populated by any agent that tracks token usage.
-type TokenUsage struct {
-	// InputTokens is the number of input tokens (fresh, not from cache)
-	InputTokens int `json:"input_tokens"`
-	// CacheCreationTokens is tokens written to cache (billable at cache write rate)
-	CacheCreationTokens int `json:"cache_creation_tokens"`
-	// CacheReadTokens is tokens read from cache (discounted rate)
-	CacheReadTokens int `json:"cache_read_tokens"`
-	// OutputTokens is the number of output tokens generated
-	OutputTokens int `json:"output_tokens"`
-	// APICallCount is the number of API calls made
-	APICallCount int `json:"api_call_count"`
-	// SubagentTokens contains token usage from spawned subagents (if any)
-	SubagentTokens *TokenUsage `json:"subagent_tokens,omitempty"`
-}
+// TokenUsage is defined in the leaf agent/types package so the checkpoint
+// contract can reference it without importing the full agent package. The
+// alias keeps existing agent.TokenUsage references working.
+type TokenUsage = types.TokenUsage
