+++ cmd/entire/cli/agent/types/token_usage.go
@@ -1,0 +1,18 @@
+package types
+
+// TokenUsage represents aggregated token usage for a checkpoint.
+// This is agent-agnostic and can be populated by any agent that tracks token usage.
+type TokenUsage struct {
+	// InputTokens is the number of input tokens (fresh, not from cache)
+	InputTokens int `json:"input_tokens"`
+	// CacheCreationTokens is tokens written to cache (billable at cache write rate)
+	CacheCreationTokens int `json:"cache_creation_tokens"`
+	// CacheReadTokens is tokens read from cache (discounted rate)
+	CacheReadTokens int `json:"cache_read_tokens"`
+	// OutputTokens is the number of output tokens generated
+	OutputTokens int `json:"output_tokens"`
+	// APICallCount is the number of API calls made
+	APICallCount int `json:"api_call_count"`
+	// SubagentTokens contains token usage from spawned subagents (if any)
+	SubagentTokens *TokenUsage `json:"subagent_tokens,omitempty"`
+}
