+++ cmd/entire/cli/checkpoint/checkpoint.go
@@ -12,7 +12,6 @@ import (
 	"errors"
 	"time"
 
-	"github.com/entireio/cli/cmd/entire/cli/agent"
 	"github.com/entireio/cli/cmd/entire/cli/agent/types"
 	"github.com/entireio/cli/cmd/entire/cli/checkpoint/id"
 	"github.com/entireio/cli/redact"
@@ -250,10 +249,10 @@ type WriteCommittedOptions struct {
 	// and the deprecated CommittedMetadata.TranscriptLinesAtStart for backward compatibility.
 
 	// TokenUsage contains the token usage for this checkpoint
-	TokenUsage *agent.TokenUsage
+	TokenUsage *types.TokenUsage
 
 	// SkillEvents records explicit native skill signals observed in this session.
-	SkillEvents []agent.SkillEvent
+	SkillEvents []types.SkillEvent
 
 	// SessionMetrics contains hook-provided session metrics (duration, turns, context usage)
 	SessionMetrics *SessionMetrics
@@ -338,7 +337,7 @@ type UpdateCommittedOptions struct {
 	Agent types.AgentType
 
 	// SkillEvents replaces the session metadata skill_events when non-empty.
-	SkillEvents []agent.SkillEvent
+	SkillEvents []types.SkillEvent
 
 	// PrecomputedBlobs, if non-nil, provides chunk blob hashes and the
 	// content-hash blob hash computed once for this transcript. When set,
@@ -469,12 +468,12 @@ type CommittedMetadata struct {
 	TranscriptLinesAtStart int `json:"transcript_lines_at_start,omitempty"`
 
 	// Token usage for this checkpoint
-	TokenUsage *agent.TokenUsage `json:"token_usage,omitempty"`
+	TokenUsage *types.TokenUsage `json:"token_usage,omitempty"`
 
 	// SkillEvents records explicit native skill signals observed in this session.
 	// Consumers use these anchors to collapse skill-related raw transcript events.
 	SkillEventsVersion int                `json:"skill_events_version,omitempty"`
-	SkillEvents        []agent.SkillEvent `json:"skill_events,omitempty"`
+	SkillEvents        []types.SkillEvent `json:"skill_events,omitempty"`
 
 	// SessionMetrics contains hook-provided session metrics (duration, turns, context usage).
 	// Populated for agents that provide these metrics via hooks (e.g., Cursor).
@@ -557,7 +556,7 @@ type CheckpointSummary struct {
 	CheckpointsCount    int                 `json:"checkpoints_count"`
 	FilesTouched        []string            `json:"files_touched"`
 	Sessions            []SessionFilePaths  `json:"sessions"`
-	TokenUsage          *agent.TokenUsage   `json:"token_usage,omitempty"`
+	TokenUsage          *types.TokenUsage   `json:"token_usage,omitempty"`
 	CombinedAttribution *InitialAttribution `json:"combined_attribution,omitempty"`
 
 	// HasReview is the umbrella "any review happened" flag: true when at least
